% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_word_sentiment}
\alias{get_word_sentiment}
\title{get_word_sentiment}
\usage{
get_word_sentiment(word)
}
\arguments{
\item{word}{character.}
}
\value{
A list of all datasets available on lexiconPT filtered by the inputed word.
}
\description{
Lookup word or term in datasets available on lexiconPT
}
\examples{
{
  get_word_sentiment("cantar")
  \dontshow{
  get_word_sentiment("facil")
  }
 }
}
