% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oplexicon_v2.1}
\alias{oplexicon_v2.1}
\title{OpLexicon V2.1}
\format{A data frame with 30677 rows and 3 variables:
\describe{
  \item{\code{term}}{character. The word or term.}
  \item{\code{type}}{character. Grammar classification of the word or term.}
  \item{\code{polarity}}{integer. Numeric classification of the polarity or sentiment. It can only assume the values of -1, 0, and 1}
}}
\source{
\url{http://ontolp.inf.pucrs.br/Recursos/downloads-OpLexicon.php}
}
\usage{
oplexicon_v2.1
}
\description{
OpLexicon is a sentiment lexicon for the Portuguese language. Please see SOUZA and VIEIRA (2012) and
SOUZA et al. (2012) for its complete reference and documentation.
}
\keyword{datasets}
