% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oplexicon_v3.0}
\alias{oplexicon_v3.0}
\title{OpLexicon V3.0}
\format{A data frame with 32191 rows and 4 variables:
\describe{
  \item{\code{term}}{character. The word of term. It also includes emoticons.}
  \item{\code{type}}{character. The type of the term.}
  \item{\code{polarity}}{integer. Numeric classification of the polarity or sentiment.}
  \item{\code{polarity_revision}}{character. Was the polarity obtained manually (A) or automatically (C)?}
}}
\source{
\url{http://ontolp.inf.pucrs.br/Recursos/downloads-OpLexicon.php}
}
\usage{
oplexicon_v3.0
}
\description{
OpLexicon is a sentiment lexicon for the Portuguese language. Please see SOUZA and VIEIRA (2012) and
SOUZA et al. (2012) for its complete reference and documentation.
}
\keyword{datasets}
