% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sentiLex_lem_PT02}
\alias{sentiLex_lem_PT02}
\title{SentiLex-PT02}
\format{A data frame with 7014 rows and 5 variables:
\describe{
  \item{\code{term}}{character The word of term.}
  \item{\code{grammar_category}}{character. The grammar classification of the term.}
  \item{\code{polarity}}{double. Numeric classification of the polarity or sentiment.}
  \item{\code{polarity_target}}{character. Polarity target. It can be N0 (subject), N1 (complement) or N2 (no documentation was found for what it means).}
  \item{\code{polarity_classification}}{character. Was the polarity obtained manually (MAN) or automatically (JALC)?.}
}}
\source{
Mário J. Silva, Paula Carvalho, Carlos Costa, Luís Sarmento, Automatic Expansion of a Social Judgment Lexicon for Sentiment Analysis Technical Report. TR 10-08. University of Lisbon, Faculty of Sciences, LASIGE, December 2010. doi: 10455/6694
}
\usage{
sentiLex_lem_PT02
}
\description{
A sentiment lexicon designed for the extraction of sentiment and opinion about human entities in Portuguese texts.
Please see SILVA, CARVALHO, COSTA and SARMENTO (2010) for its complete reference and documentation.
}
\keyword{datasets}
